Create benchmarks from `PCAWideDatasetsTests`
